Move TransactionConfig into atlasdb-impl-shared so it can be referenced in this project
